应用接入 应用列表

+++ application/admin/controller/AppGroup.php
@@ -2,6 +2,7 @@
 
 namespace app\admin\controller;
 
+use app\model\AdminApp;
 use app\model\AdminAppGroup;
 use app\util\ReturnCode;
 
@@ -74,8 +75,20 @@ class AppGroup extends Base {
         if(!$hash) {
             return $this->buildFailed(ReturnCode::EMPTY_PARAMS, '缺少必要参数');
         }
+        $has = (new AdminApp())->where(['app_group' => $hash])->count();
+        if($has) {
+            return $this->buildFailed(ReturnCode::EMPTY_PARAMS, '当前分组存在' . $has . '个应用，禁止删除');
+        }
         AdminAppGroup::destroy(['hash' => $hash]);
 
         return $this->buildSuccess();
+    }
+
+    public function getAll() {
+        $listInfo = (new AdminAppGroup())->where(['status' => 1])->select();
+
+        return $this->buildSuccess([
+            'list' => $listInfo,
+        ]);
     }
 }
